Updated Index html
You're calling javascript file first in your head. So at the time of its execution, the <a> tag doesn't exist.
Solution 1: include the script file at the bottom of the page, so by the time it executes, the rest of the DOM has been loaded.
Solution 2: call document onload method. Or use jquery $(document).ready()